@@ -1,23 +1,44 @@
 "use client";
 
-// Import the React module
 import React, { useState, useEffect } from "react";
-
-// Import the Firebase modules
-import { ref, uploadBytes, getDownloadURL, getStorage , deleteObject } from "firebase/storage";
+import {
+  ref,
+  uploadBytes,
+  getDownloadURL,
+  getStorage,
+  deleteObject,
+} from "firebase/storage";
 import { db } from "./firebase";
-import {collection,addDoc,deleteDoc,getDocs,query,onSnapshot,} from "firebase/firestore";
-
-// Import the UI components
+import {
+  collection,
+  addDoc,
+  deleteDoc,
+  query,
+  onSnapshot,
+  doc,
+  updateDoc,
+} from "firebase/firestore";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import {  Card,  CardContent,  CardFooter,  CardHeader, CardTitle} from "@/components/ui/card";
-import { Dialog,  DialogTrigger} from "@/components/ui/dialog";
-import {  Plus, Trash2, Edit, ChefHat } from "lucide-react";
+import {
+  Card,
+  CardContent,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import { Plus, Trash2, Edit, ChefHat } from "lucide-react";
 
-// Define the type for the items
 interface Item {
+  id: string;
   name: string;
   quantity: number;
   unit: string;
@@ -25,60 +46,53 @@ interface Item {
 }
 
 export default function Home() {
-
-  // Items state
   const [items, setItems] = useState<Item[]>([]);
-
-  // Item state
-  const [newItem, setNewItem] = useState<Item>({
+  const [newItem, setNewItem] = useState<Omit<Item, "id">>({
     name: "",
     quantity: 0,
     unit: "",
     imageUrl: "",
   });
+  const [searchQuery, setSearchQuery] = useState("");
+  const [filteredItems, setFilteredItems] = useState<Item[]>([]);
+  const [imageFile, setImageFile] = useState<File | null>(null);
+  const [editingItem, setEditingItem] = useState<Item | null>(null);
+  const [isDialogOpen, setIsDialogOpen] = useState(false);
 
-  // Search state
-const [searchQuery, setSearchQuery] = useState("");
-const [filteredItems, setFilteredItems] = useState<Item[]>([]);
+  const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const query = e.target.value;
+    setSearchQuery(query);
+    const filtered = items.filter((item) =>
+      item.name.toLowerCase().includes(query.toLowerCase())
+    );
+    setFilteredItems(filtered);
+  };
 
-const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-  const query = e.target.value;
-  setSearchQuery(query);
-  const filtered = items.filter((item) => item.name.toLowerCase().includes(query.toLowerCase()));
-  setFilteredItems(filtered);
-}
-  
-  // Image state
-  const [imageFile, setImageFile] = useState<File | null>(null)
-
-  // Handle file upload
   const handleFileUpload = async (file: File) => {
     const storage = getStorage();
     const storageRef = ref(storage, `images/${file.name}`);
     await uploadBytes(storageRef, file);
-    const downloadURL = await getDownloadURL(storageRef);
-    return downloadURL;
-  }
+    return getDownloadURL(storageRef);
+  };
 
-// Add item to db
-const addItem = async (e: React.FormEvent) => {
-  e.preventDefault();
-  if (newItem.name && imageFile) {
-    const imageUrl = await handleFileUpload(imageFile);
-    await addDoc(collection(db, "items"), {
-      ...newItem, imageUrl
-    });
-    setNewItem({
-      name: "",
-      quantity: 0,
-      unit: "",
-      imageUrl: "",
-    })
-    setImageFile(null);
-  }
-};
+  const addItem = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (newItem.name && imageFile) {
+      const imageUrl = await handleFileUpload(imageFile);
+      await addDoc(collection(db, "items"), {
+        ...newItem,
+        imageUrl,
+      });
+      setNewItem({
+        name: "",
+        quantity: 0,
+        unit: "",
+        imageUrl: "",
+      });
+      setImageFile(null);
+    }
+  };
 
-  // Fetch items from db
   useEffect(() => {
     const q = query(collection(db, "items"));
     const unsubscribe = onSnapshot(q, (querySnapshot) => {
@@ -86,6 +100,7 @@ const addItem = async (e: React.FormEvent) => {
       querySnapshot.forEach((doc) => {
         const data = doc.data();
         itemsArr.push({
+          id: doc.id,
           name: data.name,
           quantity: data.quantity,
           unit: data.unit,
@@ -98,107 +113,146 @@ const addItem = async (e: React.FormEvent) => {
     return () => unsubscribe();
   }, []);
 
-  //delete item from db
-  const deleteItem = async (name: string) => {
-    const q = query(collection(db, "items"));
-    const querySnapshot = await getDocs(q);
-    querySnapshot.forEach(async (doc) => {
-      if (doc.data().name === name) {
-        const data = doc.data();
-        const imageUrl = data.imageUrl
+  const deleteItem = async (id: string, imageUrl: string) => {
+    await deleteDoc(doc(db, "items", id));
+    if (imageUrl) {
+      const storage = getStorage();
+      const url = new URL(imageUrl);
+      const imagePath = decodeURIComponent(url.pathname.split("/o/")[1]);
+      const imageRef = ref(storage, imagePath);
+      await deleteObject(imageRef);
+    }
+    setItems((prevItems) => prevItems.filter((item) => item.id !== id));
+    setFilteredItems((prevFilteredItems) =>
+      prevFilteredItems.filter((item) => item.id !== id)
+    );
+  };
 
-        await deleteDoc(doc.ref);
+  const startEditing = (item: Item) => {
+    setEditingItem(item);
+    setIsDialogOpen(true);
+  };
 
-        if (imageUrl) {
-          const storage = getStorage()
-          const url = new URL(imageUrl)
-          const imagePath = decodeURIComponent(url.pathname.split('/o/')[1])
-          const imageRef = ref(storage, imagePath)
-          await deleteObject(imageRef)
-        } 
+  const handleEditChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (editingItem) {
+      setEditingItem({
+        ...editingItem,
+        [e.target.name]:
+          e.target.name === "quantity"
+            ? Number(e.target.value)
+            : e.target.value,
+      });
+    }
+  };
 
-        // Update the state
-        setItems(prevItems => prevItems.filter(item => item.name !== name))
-        setFilteredItems(prevFilteredItems => prevFilteredItems.filter(item => item.name !== name))
-      }
-    })
+  const saveEdit = async () => {
+    if (editingItem) {
+      await updateDoc(doc(db, "items", editingItem.id), {
+        name: editingItem.name,
+        quantity: editingItem.quantity,
+        unit: editingItem.unit,
+      });
+      setEditingItem(null);
+      setIsDialogOpen(false);
+    }
+  };
 
-};
-
-  // Return the UI
   return (
     <main>
       <div className="container mx-auto p-4">
-        <h1 className="text-3x1 font-bold mb-4">Pantry Tracker</h1>
+        <h1 className="text-3xl font-bold mb-4">Pantry Tracker</h1>
         <div className="mb-4">
-          <input
+          <Input
             type="text"
-            placeholder="search items..."
+            placeholder="Search items..."
             className="w-full"
             value={searchQuery}
             onChange={handleSearch}
           />
         </div>
 
-        {searchQuery && (
-          <div className="grid grid-cols-2 md:grid-cols-1 lg:grid-cols-3 gap-4 mb-4">
-            {filteredItems.map((item, index) => (
-              <Card key={index}>
-                <CardHeader>
-                  <CardTitle>{item.name}</CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <img src={item.imageUrl} alt={item.name} />
-                  <p>
-                    {item.quantity} {item.unit}
-                  </p>
-                </CardContent>
-                <CardFooter className="flex justify-between">
-                  <Button variant="outline">
-                    <Edit className="w-4 h04 mr-2"></Edit>
-                  </Button>
-                  <Button
-                    onClick={() => {
-                      deleteItem(item.name);
-                    }}
-                    variant="destructive"
-                  >
-                    <Trash2 className="w-4 h-4 mr-2">Delete</Trash2>
-                  </Button>
-                </CardFooter>
-              </Card>
-            ))}
-          </div>
-        )}
-
-        <div className="grid grid-cols-2 md:grid-cols-1 lg:grid-cols-3 gap-4 mb-4">
-          {items.map((item, index) => (
-            <Card key={index}>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mb-4">
+          {(searchQuery ? filteredItems : items).map((item) => (
+            <Card key={item.id}>
               <CardHeader>
                 <CardTitle>{item.name}</CardTitle>
               </CardHeader>
               <CardContent>
-                <img src={item.imageUrl} alt={item.name} />
+                <img
+                  src={item.imageUrl}
+                  alt={item.name}
+                  className="w-full h-48 object-cover mb-2"
+                />
                 <p>
                   {item.quantity} {item.unit}
                 </p>
               </CardContent>
               <CardFooter className="flex justify-between">
-                <Button variant="outline">
-                  <Edit className="w-4 h04 mr-2"></Edit>
+                <Button variant="outline" onClick={() => startEditing(item)}>
+                  <Edit className="w-4 h-4 mr-2" />
+                  Edit
                 </Button>
                 <Button
-                  onClick={() => {
-                    deleteItem(item.name);
-                  }}
+                  onClick={() => deleteItem(item.id, item.imageUrl)}
                   variant="destructive"
                 >
-                  <Trash2 className="w-4 h-4 mr-2">Delete</Trash2>
+                  <Trash2 className="w-4 h-4 mr-2" />
+                  Delete
                 </Button>
               </CardFooter>
             </Card>
           ))}
         </div>
+
+        <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
+          <DialogContent>
+            <DialogHeader>
+              <DialogTitle>Edit Item</DialogTitle>
+            </DialogHeader>
+            {editingItem && (
+              <div className="grid gap-4 py-4">
+                <div className="grid grid-cols-4 items-center gap-4">
+                  <Label htmlFor="name" className="text-right">
+                    Name
+                  </Label>
+                  <Input
+                    id="name"
+                    name="name"
+                    value={editingItem.name}
+                    onChange={handleEditChange}
+                    className="col-span-3"
+                  />
+                </div>
+                <div className="grid grid-cols-4 items-center gap-4">
+                  <Label htmlFor="quantity" className="text-right">
+                    Quantity
+                  </Label>
+                  <Input
+                    id="quantity"
+                    name="quantity"
+                    type="number"
+                    value={editingItem.quantity}
+                    onChange={handleEditChange}
+                    className="col-span-3"
+                  />
+                </div>
+                <div className="grid grid-cols-4 items-center gap-4">
+                  <Label htmlFor="unit" className="text-right">
+                    Unit
+                  </Label>
+                  <Input
+                    id="unit"
+                    name="unit"
+                    value={editingItem.unit}
+                    onChange={handleEditChange}
+                    className="col-span-3"
+                  />
+                </div>
+              </div>
+            )}
+            <Button onClick={saveEdit}>Save Changes</Button>
+          </DialogContent>
+        </Dialog>
 
         <Card>
           <CardHeader>
@@ -212,48 +266,45 @@ const addItem = async (e: React.FormEvent) => {
                   id="name"
                   placeholder="Enter item name"
                   value={newItem.name}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                    setNewItem({ ...newItem, name: e.target.value });
-                  }}
-                ></Input>
+                  onChange={(e) =>
+                    setNewItem({ ...newItem, name: e.target.value })
+                  }
+                />
               </div>
               <div className="flex flex-col space-y-1.5">
                 <Label htmlFor="quantity">Quantity</Label>
                 <Input
-                  value={newItem.quantity}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                    setNewItem({
-                      ...newItem,
-                      quantity: parseInt(e.target.value),
-                    });
-                  }}
                   id="quantity"
                   type="number"
                   placeholder="Enter quantity"
+                  value={newItem.quantity}
+                  onChange={(e) =>
+                    setNewItem({ ...newItem, quantity: Number(e.target.value) })
+                  }
                 />
               </div>
               <div className="flex flex-col space-y-1.5">
                 <Label htmlFor="unit">Unit</Label>
                 <Input
-                  value={newItem.unit}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-                    setNewItem({ ...newItem, unit: e.target.value });
-                  }}
                   id="unit"
                   placeholder="Enter unit (e.g., boxes, cans)"
+                  value={newItem.unit}
+                  onChange={(e) =>
+                    setNewItem({ ...newItem, unit: e.target.value })
+                  }
                 />
               </div>
               <div className="flex flex-col space-y-1.5">
                 <Label htmlFor="image">Image</Label>
                 <Input
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+                  id="image"
+                  type="file"
+                  onChange={(e) => {
                     if (e.target.files && e.target.files[0]) {
                       setImageFile(e.target.files[0]);
                     }
                   }}
-                  id="image"
-                  type="file"
-                ></Input>
+                />
               </div>
             </div>
           </CardContent>
@@ -272,7 +323,9 @@ const addItem = async (e: React.FormEvent) => {
               Suggest Recipe
             </Button>
           </DialogTrigger>
-          <p>Here is a recipe suggestion based on your pantry items...</p>
+          <DialogContent>
+            <p>Here is a recipe suggestion based on your pantry items...</p>
+          </DialogContent>
         </Dialog>
       </div>
     </main>
